Add files via upload
In this scenario, for a burfi of size 20x25, a better profit value of 195k can be achieved. This does not prove
that 195k is an optimal distribution, as keeping the burfi size 20x25 will yield a better profit value of 255k.
The primary goal of the problem is to ascertain the optimal solution that generates the maximum profit
through any distribution method. The algorithm should explore different splits to achieve the maximum
profit without requiring the specific distribution details.
Tasks to be accomplished:
1. Develop a Simple Recursive Top-Down algorithm without Memoization.